pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Hexlet project - Page Analyzer"
authors = ["DanilMullakhmetov <danilmullakhmetov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.0"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pytest-cov = "^5.0.0"
flake8 = "^7.1.1"  # Оставьте только одну версию flake8

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
